Finish Data-base Back End & Pseudo-Interface
This is the complete program. Copy it to your interpreter and let me know what I need to change/improve. Go crazy and test it for errors.
Don't hit the "VIEW TRANSACTIONS" option. We haven't developed code for that part yet. That's where the pie charts and graphs and all the other stuff are going to come. Cheers!