Added User Window
Hey Ayush I added the class definition for the user window.
The problem is the frames are not completely overlapping each other (run the program and you'll understand) so I need your help to figure out what to do. 
Also, when I meant verify the email, I wanted you to add the verify_email function in our CLI (Command Line Interface) program and use that result. This means you'll have to add a label and an entry and another submit button for the user to enter the verification code sent by the program. 
To add these extra widgets in the frame, you'll have to add some more rows and columns and then grid them accordingly. Don't worry - I've changed the code so that it's much easier to visualize the layout. 

Cheers
Joshua